首页增加小班课信息

+++ boxfish-online-card/src/main/java/com/boxfishedu/workorder/web/view/fishcard/FishCardGroupsInfo.java
@@ -12,7 +12,7 @@ public class FishCardGroupsInfo {
 
     private Long workOrderId;// 鱼卡id
 
-    private Integer chatRoomId;//房间号
+    private Long chatRoomId;//房间号
 
     private String [] memberAccount;//学生id 老师id
